started adding items

+++ mouse_control.h
@@ -16,7 +16,7 @@ namespace Mouse {
 	int Mouse_Selection_Box_x_Display;
 	int Mouse_Selection_Box_y_Display;
 
-	bool Inside_Cursor(int aX, int aY, float bX, float bY, float squareRadius) { //a is mouse, b is target of the mouse
+	bool Inside_Cursor(float aX, float aY, float bX, float bY, float squareRadius) { //a is mouse, b is target of the mouse
 		if (aX > bX - squareRadius && aX < bX + squareRadius && aY < bY + squareRadius && aY > bY - squareRadius) {
 			return true;
 		}
